[TASK] Use proven selenium version used by TYPO3 core
A newer selenium image has been used, which seems to
introduces further issues. In core-development, this
is known to be very fragile, therefor it is seldom
changed.

This change uses the same version like the TYPO3 core,
which has been proven to be stable for now. Additionally,
the arch switch for ARM based image is introduced.